requirements.txt
agate==1.7.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickhouse-connect==0.8.15
clickhouse-driver==0.2.9
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.14.1
dbt-clickhouse==1.8.9
dbt-common==1.16.0
dbt-core==1.8.9
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
deepdiff==7.0.1
future==1.0.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
hologram==0.0.16
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
Logbook==1.5.3
lz4==4.4.3
MarkupSafe==3.0.2
mashumaro==3.14
minimal-snowplow-tracker==0.0.2
more-itertools==10.6.0
msgpack==1.1.0
networkx==2.8.8
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.2
parsedatetime==2.4
pathspec==0.11.2
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.8.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==2.3.8
zipp==3.21.0
zstandard==0.23.0
